requirements.txt
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
